python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiortc>=1.13.0",
    "boto3>=1.39.4",
    "deepgram-sdk>=5.2.0",
    "fastapi[standard]>=0.115.13",
    "google-cloud-texttospeech>=2.31.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-redis>=0.1.1",
    "numpy>=1.26.0",
    "lxml>=6.0.0",
    "neo4j>=5.28.2",
    "openai>=1.88.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
    "livekit>=0.17.5",
    "livekit-api>=1.0.7",
    "copilotkit>=0.0.1",
    "celery>=5.6.0",
    "supabase>=2.24.0",
    "google-genai>=1.56.0",
    "pdf2image>=1.17.0",
    "langchain-google-genai>=2.0.0",
    "pymupdf>=1.25.0",
    "tavily-python>=0.5.0",
    # Lightweight document extractors (replace Unstructured for DOCX/PPTX)
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "fakeredis>=2.20.0",
    "httpx>=0.27.0",
]
# Legacy fallback - only install if you need Unstructured's advanced parsing
# Includes Tesseract OCR dependency (requires system package: tesseract-ocr)
legacy = [
    "unstructured[all-docs]>=0.18.3",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Explicitly specify packages for flat layout
py-modules = []

[tool.setuptools.packages.find]
include = ["lib*", "agents*", "tasks*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "*.pyc", "__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]

